로그아웃 로그인 권한 체크

+++ src/main/java/com/KoreaIT/demo/controller/UsrArticleController.java
@@ -11,6 +11,7 @@ import com.KoreaIT.demo.service.ArticleService;
 import com.KoreaIT.demo.util.Util;
 import com.KoreaIT.demo.vo.Article;
 import com.KoreaIT.demo.vo.ResultData;
+import javax.servlet.http.HttpSession;
 
 @Controller
 public class UsrArticleController {
@@ -25,21 +26,25 @@ public class UsrArticleController {
 	// 액션 메서드
 	@RequestMapping("/usr/article/doAdd")
 	@ResponseBody
-	public ResultData<Article> doAdd(String title, String body) {
+	public ResultData<Article> doAdd(HttpSession httpSession, String title, String body) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
 		
 		if (Util.empty(title)) {
-			return ResultData.from("F-1","제목을 입력해주세요");
+			return ResultData.from("F-1", "제목을 입력해주세요");
 		}
 		
 		if (Util.empty(body)) {
-			return ResultData.from("F-2","내용을 입력해주세요");
+			return ResultData.from("F-2", "내용을 입력해주세요");
 		}
 		
-		articleService.writeArticle(title, body);
+		articleService.writeArticle((int) httpSession.getAttribute("loginedMemberId"), title, body);
 		
 		int id = articleService.getLastInsertId();
 		
-		return ResultData.from("S-1",Util.f("%d번 게시물이 생성 되었습니다.", id), articleService.getArticleById(id));
+		return ResultData.from("S-1", Util.f("%d번 게시물이 생성되었습니다", id), articleService.getArticleById(id));
 	}
 	
 	@RequestMapping("/usr/article/getArticle")
@@ -49,8 +54,7 @@ public class UsrArticleController {
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
-//			return id + "번 게시물은 존재하지 않습니다.";
-			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다.", id));
+			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 		
 		return ResultData.from("S-1", Util.f("%d번 게시물 입니다", id), article);
@@ -58,37 +62,46 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/getArticles")
 	@ResponseBody
-	public List<Article> getArticles() {
-		return articleService.getArticles();
+	public ResultData<List<Article>> getArticles() {
+		
+		return ResultData.from("S-1", "게시물 리스트", articleService.getArticles());
 	}
 	
 	@RequestMapping("/usr/article/doModify")
 	@ResponseBody
-	public String doModify(int id, String title, String body) {
+	public ResultData doModify(HttpSession httpSession, int id, String title, String body) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
 		
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
-			return id + "번 게시물은 존재하지 않습니다.";
+			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 		
 		articleService.modifyArticle(id, title, body);
 		
-		return id + "번 게시물을 수정했습니다";
+		return ResultData.from("S-1", Util.f("%d번 게시물을 수정했습니다", id));
 	}
 	
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public String doDelete(int id) {
+	public ResultData doDelete(HttpSession httpSession, int id) {
+		
+		if (httpSession.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
 		
 		Article article = articleService.getArticleById(id);
 		
 		if(article == null) {
-			return id + "번 게시물은 존재하지 않습니다.";
+			return ResultData.from("F-1", Util.f("%d번 게시물은 존재하지 않습니다", id));
 		}
 		
 		articleService.deleteArticle(id);
 		
-		return id + "번 게시물을 삭제했습니다";
+		return ResultData.from("S-1", Util.f("%d번 게시물을 삭제했습니다", id));
 	}
 }
